pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "debugagent"
version = "0.1.0"
description = "Self-improving autonomous debugging agent"
requires-python = ">=3.11"
dependencies = [
    "google-genai",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.7.0",
    "pydantic>=2.0.0",
    "typer[all]>=0.12.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "langfuse>=2.0.0",
    "pytest>=8.0.0",
    "pytest-timeout>=2.3.0",
    "pytest-json-report>=1.5.0",
]

[project.optional-dependencies]
dev = ["black", "isort", "pytest-cov"]

[project.scripts]
debugagent = "debugagent.main:app"

[tool.setuptools.packages.find]
include = ["debugagent*"]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
